feat: shared db redaction
