+++ db/core/Cargo.toml
@@ -37,6 +37,7 @@ sqlx = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
+url = { workspace = true }
 validator = { workspace = true }
 
 blokli-db-entity = { workspace = true }
